refactor(two-factor-test): Move two-factor integration tests into a CODEOWNERS-compatible path.
